Resolve issue preventing search index being built
Currently there is a problem in which the search index
(search_index.json) used by `lunr` is not being built

This problem is the same as one earlier faced n which the antora command
is being incorrectly invoked, using the:

 antora --require generator ${PLAYBOOK}

instead of the

  antora --generator generator ${PLAYBOOK}

syntax expected. This means there is no invocation of the index building
scripts, and no index built.

This commit corrects that command.