Override the header with less content
By default the Antora documentation supplies a set of links in the
header partial that are not connected to anything else. These look a
little ridiculous on user facing documentation, and should thus be
removed.

This commit does so by introducing the Antora partials configuration and
overriding the header partial with an empty one.