Add the blog taxonomy
There is a set of publications that are designed to help those who are
interested in software development, technology and littleman.co
understand how each of those topics work. However, the topics are
somewhat disparate; designed for different audiences.

This commit adds a taxonomy describing the anticipated types of
documentation, as well as where each documentation should be published.

It's hoped this will allow users to stay in contact only in the sort of
content they're interested in reading about.

Resolved #3