Add development tools
One of the primary costs in software development is becoming familiar
with the various tools used to track and manage that software over time.
In order to minimize the time required to learn the new tools this
commit publishes the tools that are used at littleman.co

The hope is that those who are interested in hiring littleman.co's
services will also likely be those who use the tools.