Add a helm chart for deployment
Currently the deployment is deployed on Netlify. While Netlify is cool,
the goal is to use this as a bit of a demonstration of the appropriate
technology, surfacing the various tools and processes that we can use.

Netlify is not sufficiently useful to adopt in all cases; tooling is
thus cheaper over time specialised in something that can (k8s + GKE)