Install npm resources in antora container
Antora is not installed with the base image, but rather installed as
part of the repository. However, there is currently no step in the build
environment that pulls in these dependencies.

This commit adds the command "npm install" to the antora container where
it should pull in required dependencies before generating the site.