Combine the two compilation build steps
Currently there is a problem in which content build as part of a build
step does not persist to the next build step.

The suggested solution is to push this content to a third party
repository. However, that means there will be a useless artifact in that
repository for at least a short period of time, which isn't what we
should do through the life of the build.

This commit attempts to address that by combining two build steps, in
the hope they'll be executed in the same environment sequentially.

See https://github.com/drone/drone/issues/966