Add istio configuration to make public
While there is a set of documentation hosted on the cluster now, that
documentation isn't currently visible to the public. To make it visible
it needs to be made accessible by an ingress; a client side load
balancer.

This commit introduces the appropriate configuration such that the
service can be made available via the istio mesh network, and associated
ingress.

== Design Notes ==

=== Why not non-istio ingress ===

There are currently two mechanisms for accessing the cluster:

- Istio
- NGINX

This commit uses the Istio configuration. There are a couple of
advantages to istio over NGINX:

- mTLS on all traffic
- Visibility on all traffic
- More experience with Istio