Force pod rotation by adding release time to spec
While the container for newer documentation is being correctly build and
pushed to the upstream repository, it is never being released.

The problem is severalfold:

1. There is no way to specify current release without using some
   reference through the build for this. Naively that's git refspec,
   but its also unlikely we'll ever be rolling back.
2. There's no way to force helm to update the pods, aside from
   --force-recreate which has unpredictable results.

In the end, the solution proposed is to add a field which indicates when
the release was deployed. Because this field is added to the pod
metadata field it requires a re-release of the pod.

This means that it will always use the latest container, which should
always be correct.

It additionally means that all untagged containers in the registry can
be safely removed at some point.