Add project management documentation, note bots
Currently there is no documentation that explains how projects should be
managed while on GitHub.

This commit attempts to address this by adding a project management
stub. The goal for this is to help explain what some users might see
while interacting with the GitHub issues on the project.

It makes notes that there are bots that carry out various tasks
automatically to help make things clearer in the project.

Future work should build on this to add specific examples of bots, where
they can be found and what they're doing.