Prevent the deploy running on PR
Currently there is a problem in which the deployment process runs on a
pull request against master.

This is problematic as any user can currently run this pull request,
which in principle allows any user access to the protected variables
only available on master.

This commit attempts to resolve this by limiting the master branch to
the "push" event. It's hoped that this event is triggered on a merge to
the master branch, allowing continuous delivery.

Resolves #31

== Design Notes ==

=== Duplicated docs build ===

It still makes sense to build the documentation but to discard it for
pull requests. This allows testing whether the pull request
documentation is even valid before it's pushed.